requirements.txt
flask==1.1.1
tensorflow-cpu==2.3.0
tensorflow_hub==0.10.0
matplotlib==3.3.3
numpy==1.18.0
gunicorn==20.0.4
